requirements.txt
anyio==3.6.2
APScheduler==3.10.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
kafka-python==2.0.2
pydantic==1.10.7
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.2
